Avaliações/Trabalhos/rest-app-cars-blog/api/requirements.txt
Flask==3.0.3
pyjwt==2.9.0
psycopg2-binary==2.9.10
flask-cors==5.0.0